Added the new tokens/keywords and refactored lexer into a module
